# apps/api/openapi/agent-tools.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pi Finance Agent HTTP contract",
    "version": "1.0.0"
  },
  "paths": {
...
    "/transactions/expense": {
      "post": {
        "operationId": "create_expense",
        "summary": "Register an expense",
        "x-pi-tool": {
          "label": "Create Expense",
          "idempotency": true,
          "result": {
            "kind": "first",
            "key": "transactionId",
            "path": "id"
          }
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": true,
            "x-pi-context": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "idempotencyKey",
            "in": "header",
            "x-pi-context": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "description",
                  "amountCents",
                  "categoryId",
                  "accountId",
                  "date"
                ],
                "properties": {
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 240
                  },
                  "amountCents": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "accountId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "categoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "subcategoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 2000
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/transactions/income": {
      "post": {
        "operationId": "create_income",
        "summary": "HTTP contract for create_income",
        "x-pi-tool": {
          "name": "create_income",
          "label": "create_income",
          "description": "HTTP contract for create_income",
          "method": "POST",
          "path": "/transactions/income",
          "idempotency": true,
          "shadow": false,
          "result": null
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "idempotencyKey",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 240
                  },
                  "amountCents": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "accountId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "categoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "subcategoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "required": [
                  "description",
                  "amountCents",
                  "date",
                  "accountId",
                  "categoryId"
                ]
              }
            }
          }
        }
      }
    },
...
    "/transfers": {
      "post": {
        "operationId": "create_transfer",
        "summary": "HTTP contract for create_transfer",
        "x-pi-tool": {
          "name": "create_transfer",
          "label": "create_transfer",
          "description": "HTTP contract for create_transfer",
          "method": "POST",
          "path": "/transfers",
          "idempotency": true,
          "shadow": false,
          "result": null
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "idempotencyKey",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 240
                  },
                  "amountCents": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "fromAccountId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "toAccountId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "description",
                  "amountCents",
                  "date",
                  "fromAccountId",
                  "toAccountId"
                ]
              }
            }
          }
        }
      }
    },
...
    "/transactions/{transactionId}": {
      "delete": {
        "operationId": "delete_transaction",
        "summary": "HTTP contract for delete_transaction",
        "x-pi-tool": {
          "name": "delete_transaction",
          "label": "delete_transaction",
          "description": "HTTP contract for delete_transaction",
          "method": "DELETE",
          "path": "/transactions/{transactionId}",
          "idempotency": true,
          "shadow": false,
          "result": null
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "idempotencyKey",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "transactionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "patch": {
        "operationId": "update_transaction",
        "summary": "HTTP contract for update_transaction",
        "x-pi-tool": {
          "name": "update_transaction",
          "label": "update_transaction",
          "description": "HTTP contract for update_transaction",
          "method": "PATCH",
          "path": "/transactions/{transactionId}",
          "idempotency": true,
          "shadow": false,
          "result": null
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "idempotencyKey",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "transactionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 240
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "amountCents": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "accountId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "categoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "subcategoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "required": []
              }
            }
          }
        }
      }
    },
...
  }
}